Known Error Fix
There was an error with the way that JSON data was handled. If a key value from a response had a single quote or double quote, then it would break the formatting of the JSON data by closing the value prematurely. This has been fixed by parsing through every character in the JSON data and removing and quotes that would prematurely close the key value. This may still occur in Dehash JSON data, but we have not seen a case of that, and we plan on implementing a pre-emptive check for that.